Move binary scanning into its own top-level module, improve documentation
