+++ rustsec/src/binary_scanning/binary_deps.rs
@@ -8,12 +8,12 @@ use crate::{Error, ErrorKind};
 use auditable_serde::VersionInfo;
 use cargo_lock::{Dependency, Lockfile, Package};
 
-use crate::binary_format::BinaryFormat;
+use crate::binary_scanning::BinaryFormat;
 
 /// The default file size limit is 8MB
 const DEFAULT_FILE_SIZE_LIMIT: usize = 8 * 1024 * 1024;
 
-/// Defines the state of a Lockfile
+/// Dependencies recovered from scanning a compiled Rust executable
 pub enum BinaryReport {
     /// Full dependency list embedded by `cargo auditable`
     Complete(Lockfile),
@@ -23,13 +23,18 @@ pub enum BinaryReport {
     None,
 }
 
-/// Load the dependency tree from some binary. By default the file size limit is 8MB.
+/// Load the dependency tree from a compiled Rust executable.
+///
+/// Recovers the precise dependency list if the binary is built with [`cargo auditable`](https://crates.io/crates/cargo-auditable).
+/// Failing that, recovers as many crates as possible from panic messages (using [quitters]).
+///
+/// If `audit_data_size_limit` is set to `None`, the limit defaults to 8MB.
 pub fn load_deps_from_binary(
     file_contents: &[u8],
-    file_size_limit: Option<usize>,
+    audit_data_size_limit: Option<usize>,
 ) -> crate::Result<(BinaryFormat, BinaryReport)> {
     let format = detect_format(file_contents);
-    let file_size_limit = match file_size_limit {
+    let file_size_limit = match audit_data_size_limit {
         Some(size) => size,
         None => DEFAULT_FILE_SIZE_LIMIT,
     };
